refactor: stage management
