+++ templates/chat_interface.html
@@ -37,7 +37,7 @@
            </a>
            <a href="{{ url_for('select_problem_page') }}" class="new-chat-link">Phiên Mới</a>
            <div class="status connecting">
-               <i class="fas fa-circle" id="connectionStatusIcon"></i>
+               <span id="connectionStatusIcon"></span>
                <span id="statusText">Đang kết nối...</span>
            </div>
        </header>
